requirements.txt
Flask
pillow
fpdf
docx2pdf
gunicorn